取消订单功能代码开发

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -31,5 +31,5 @@ public interface OrderMapper {
 
     // 根据ID查订单
     @Select("select * from orders where id = #{id}")
-    Orders getById(Integer id);
+    Orders getById(Long id);
 }
